code/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.1.0
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jsmin==3.0.1
msgpack==1.0.5
Pillow==9.5.0
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2023.3
redis==4.5.4
sqlparse==0.4.3